after adding manufactoring invoice

+++ app/Http/Controllers/accountController.php
@@ -3,9 +3,11 @@
 namespace App\Http\Controllers;
 
 use App\Models\Comment;
+use App\Models\Journal;
 use Illuminate\Http\Request;
 use App\Models\Account;
 use App\functions\globalFunctions;
+use Illuminate\Support\Facades\DB;
 use phpDocumentor\Reflection\PseudoTypes\NonEmptyLowercaseString;
 
 class accountController extends Controller
@@ -35,7 +37,7 @@ class accountController extends Controller
      * Store a newly created resource in storage.
      *
      * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
+     * @return \Illuminate\Http\RedirectResponse
      */
     public function store(Request $request)
     {
@@ -102,7 +104,7 @@ class accountController extends Controller
      *
      * @param  \Illuminate\Http\Request  $request
      * @param  int  $id
-     * @return \Illuminate\Http\Response
+     * @return \Illuminate\Http\RedirectResponse
      */
     public function update(Request $request, Account $account)
     {
@@ -158,28 +160,55 @@ class accountController extends Controller
      * Remove the specified resource from storage.
      *
      * @param int $id
-     * @return \Illuminate\Http\Response
+     * @return \Illuminate\Http\RedirectResponse
      */
     public function destroy(Request $request,$account_id)
     {
         if ($account_id > 0) {
-            $result = Account::withTrashed()->find($account_id)->forceDelete();
-            globalFunctions::registerUserActivityLog("deleted","account",$account_id);
+            $balance = Journal::where("first_part_id",$account_id)->selectRaw("sum(debit) - sum(credit) as balance")->first()->balance;
+            // here whe check if the account balance is zero or not , so if it is zero delete it and update all invoices whose first_part_id = id and set first_part_id = 0 to allow adding new account with the same fo previous id
+             if ($balance != 0 and $balance != null) {
+                globalFunctions::flashMessage("softDelete","account_is_not_zero","account");
+                return back();
+            }
+            try {
+                DB::beginTransaction();
+                $result = Account::withTrashed()->find($account_id)->forceDelete();
+                Journal::where("first_part_id",$account_id)->update(["first_part_id"=>0]);
+                globalFunctions::registerUserActivityLog("deleted","account",$account_id);
+                DB::commit();
+            }
+            catch (\PDOException $e){
+                DB::rollBack();
+                $result = null;
+            }
         } else if (isset($request["multi_ids"])) {
-            $ids = $request["multi_ids"];
-            $result = Account::withTrashed()->whereIn("id",$ids)->forceDelete();
-            foreach ($ids as $id){
-                globalFunctions::registerUserActivityLog("deleted","account",$id);
+            try {
+                DB::beginTransaction();
+                $ids = $request["multi_ids"];
+                $result = Account::withTrashed()->whereIn("id",$ids)->forceDelete();
+                Journal::whereIn("first_part_id",$ids)->update(["first_part_id"=>0]);
+                foreach ($ids as $id) {
+                    $balance = Journal::where("first_part_id", $id)->selectRaw("sum(debit) - sum(credit) as balance")->first()->balance;
+                    if ($balance != 0 and $balance != null) {
+                        DB::rollBack();
+                        globalFunctions::flashMessage("softDelete", "account_is_not_zero", "account");
+                        return back();
+                    }
+                }
+                DB::commit();
+                foreach ($ids as $id){
+                    globalFunctions::registerUserActivityLog("deleted","account",$id);
+                }
+            }
+            catch (\PDOException $e){
+                DB::rollBack();
+                $result = null;
             }
         } else {
             $result = null;
         }
 
-//        if ($result!=null) {
-//            session()->flash("success",__("messages.deleted_successfully",["attribute"=>__("global.account")]));
-//        }else{
-//            session()->flash("success",__("messages.not_deleted_successfully",["attribute"=>__("global.account")]));
-//        }
         globalFunctions::flashMessage("delete",$result,"account");
 
         return back();
@@ -188,7 +217,7 @@ class accountController extends Controller
     /**
      * Remove the specified resource from storage.
      *
-     * @return \Illuminate\Http\Response
+     * @return \Illuminate\Contracts\Foundation\Application|\Illuminate\Contracts\View\Factory|\Illuminate\Contracts\View\View|\Illuminate\Http\Response
      */
     public function viewRecyclebin()
     {
@@ -204,14 +233,30 @@ class accountController extends Controller
      */
     public function softDelete(Request $request,$account_id)
     {
+
         if ($account_id > 0) {
+            $balance = Journal::where("first_part_id",$account_id)->selectRaw("sum(debit) - sum(credit) as balance")->first()->balance;
+            if ($balance != 0 and $balance != null) {
+                globalFunctions::flashMessage("softDelete","account_is_not_zero","account");
+                return back();
+            }
             $result = Account::find($account_id)->delete();
             globalFunctions::registerUserActivityLog("recycled","account",$account_id);
         } else if (isset($request["multi_ids"])) {
+            DB::beginTransaction();
             $ids = $request["multi_ids"];
-            $result = Account::whereIn("id",$ids)->delete();
-            foreach ($ids as $id){
-                globalFunctions::registerUserActivityLog("recycled","account",$id);
+            $result = Account::whereIn("id", $ids)->delete();
+            foreach ($ids as $id) {
+                $balance = Journal::where("first_part_id", $id)->selectRaw("sum(debit) - sum(credit) as balance")->first()->balance;
+                if ($balance != 0 and $balance != null) {
+                    DB::rollBack();
+                    globalFunctions::flashMessage("softDelete", "account_is_not_zero", "account");
+                    return back();
+                }
+            }
+            DB::commit();
+            foreach ($ids as $id) {
+                globalFunctions::registerUserActivityLog("recycled", "account", $id);
             }
         } else {
             $result = null;
@@ -246,7 +291,6 @@ class accountController extends Controller
             }
         } else {
             $result = null;
-            header('Location:'+$_SERVER['HTTP_REFERER']);
         }
 //        if ($result!=null) {
 //            session()->flash("success",__("messages.restored_successfully",["attribute"=>__("global.account")]));
